requirements.txt
python-dotenv
streamlit
google-generativeai
langchain
langchain-community
pypdf2
faiss-cpu